clarify file names and remove private dir
