+++ .gitignore
@@ -4,3 +4,4 @@ server/bin/
 app/doc/api
 tmp/
 data/
+private/
