--- private/data/mongodb/WiredTiger
@@ -1,2 +1,0 @@
-WiredTiger
-WiredTiger 10.0.1: (April 12, 2021)
