--- server/conf/app.ini
@@ -1,9 +1,0 @@
-[server]
-Address = 8000
-ReadTimeout = 60
-WriteTimeout = 60
-MaxHeaderBytes = 1 << 20
-
-[mongodb]
-Address = 27017
-Uri = mongodb
