+++ server/conf/conf_dev.ini
@@ -1,0 +1,8 @@
+[server]
+Address = 8000
+ReadTimeout = 60
+WriteTimeout = 60
+MaxHeaderBytes = 1 << 20
+
+[mongodb]
+Uri = mongodb://localhost:27017
